# docs/api/openapi.yaml
openapi: 3.0.0
info:
  title: Stock Analysis Platform API
  version: 1.0.0
  description: RESTful API for stock analysis and portfolio management
  contact:
    name: API Support
    email: support@example.com
servers:
  -
    url: "http://localhost:3000"
    description: Development server
  -
    url: "https://api.example.com"
    description: Production server
paths:
...
  /api/indices:
    get:
      summary: Market Routes
      description: null
      tags:
        - Market
      responses:
        200:
          description: Successful response
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
                  message:
                    type: string
              examples:
                default:
                  summary: Example success response
                  value:
                    success: true
                    data:
                    message: Operation completed successfully
        400:
          description: Bad request
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  error:
                    type: string
                  message:
                    type: string
                  details:
                    type: array
                    items:
                      type: string
              examples:
                default:
                  summary: Example error response
                  value:
                    success: false
                    error: Bad Request
                    message: Invalid request parameters
                    details:
                      - Field validation failed
        500:
          description: Internal server error
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  error:
                    type: string
                  message:
                    type: string
              examples:
                default:
                  summary: Example server error response
                  value:
                    success: false
                    error: Internal Server Error
                    message: An unexpected error occurred
  /api/breadth:
    get:
      summary: Get market breadth data (advance/decline statistics)
      description: "Implements Requirement 18.4: Display advance/decline counts (实现需求: 18.4)"
      tags:
        - Market
      responses:
        200:
          description: Successful response
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
                  message:
                    type: string
              examples:
                default:
                  summary: Example success response
                  value:
                    success: true
                    data:
                    message: Operation completed successfully
        400:
          description: Bad request
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  error:
                    type: string
                  message:
                    type: string
                  details:
                    type: array
                    items:
                      type: string
              examples:
                default:
                  summary: Example error response
                  value:
                    success: false
                    error: Bad Request
                    message: Invalid request parameters
                    details:
                      - Field validation failed
        500:
          description: Internal server error
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  error:
                    type: string
                  message:
                    type: string
              examples:
                default:
                  summary: Example server error response
                  value:
                    success: false
                    error: Internal Server Error
                    message: An unexpected error occurred
  /api/sentiment:
    get:
      summary: Get market sentiment indicators
      description: "Implements Requirement 18.4: Display market sentiment indicators (实现需求: 18.4)"
      tags:
        - Market
      responses:
        200:
          description: Successful response
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
                  message:
                    type: string
              examples:
                default:
                  summary: Example success response
                  value:
                    success: true
                    data:
                    message: Operation completed successfully
        400:
          description: Bad request
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  error:
                    type: string
                  message:
                    type: string
                  details:
                    type: array
                    items:
                      type: string
              examples:
                default:
                  summary: Example error response
                  value:
                    success: false
                    error: Bad Request
                    message: Invalid request parameters
                    details:
                      - Field validation failed
        500:
          description: Internal server error
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  error:
                    type: string
                  message:
                    type: string
              examples:
                default:
                  summary: Example server error response
                  value:
                    success: false
                    error: Internal Server Error
                    message: An unexpected error occurred
  /api/gainers:
    get:
      summary: Get top gaining stocks
      description: "Implements Requirement 18.5: Display top gainers Query Parameters: - limit: Maximum number of stocks (optional, default: 10, max: 50) (实现需求: 18.5)"
      tags:
        - Market
      responses:
        200:
          description: Successful response
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
                  message:
                    type: string
              examples:
                default:
                  summary: Example success response
                  value:
                    success: true
                    data:
                    message: Operation completed successfully
        400:
          description: Bad request
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  error:
                    type: string
                  message:
                    type: string
                  details:
                    type: array
                    items:
                      type: string
              examples:
                default:
                  summary: Example error response
                  value:
                    success: false
                    error: Bad Request
                    message: Invalid request parameters
                    details:
                      - Field validation failed
        500:
          description: Internal server error
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  error:
                    type: string
                  message:
                    type: string
              examples:
                default:
                  summary: Example server error response
                  value:
                    success: false
                    error: Internal Server Error
                    message: An unexpected error occurred
      parameters:
        -
          name: limit
          in: query
          description: Maximum number of stocks
          required: false
          schema:
            type: string
  /api/losers:
    get:
      summary: Get top losing stocks
      description: "Implements Requirement 18.5: Display top losers Query Parameters: - limit: Maximum number of stocks (optional, default: 10, max: 50) (实现需求: 18.5)"
      tags:
        - Market
      responses:
        200:
          description: Successful response
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
                  message:
                    type: string
              examples:
                default:
                  summary: Example success response
                  value:
                    success: true
                    data:
                    message: Operation completed successfully
        400:
          description: Bad request
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  error:
                    type: string
                  message:
                    type: string
                  details:
                    type: array
                    items:
                      type: string
              examples:
                default:
                  summary: Example error response
                  value:
                    success: false
                    error: Bad Request
                    message: Invalid request parameters
                    details:
                      - Field validation failed
        500:
          description: Internal server error
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  error:
                    type: string
                  message:
                    type: string
              examples:
                default:
                  summary: Example server error response
                  value:
                    success: false
                    error: Internal Server Error
                    message: An unexpected error occurred
      parameters:
        -
          name: limit
          in: query
          description: Maximum number of stocks
          required: false
          schema:
            type: string
  /api/most-active:
    get:
      summary: Get most active stocks by volume
      description: "Implements Requirement 18.5: Display volume leaders Query Parameters: - limit: Maximum number of stocks (optional, default: 10, max: 50) (实现需求: 18.5)"
      tags:
        - Market
      responses:
        200:
          description: Successful response
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
                  message:
                    type: string
              examples:
                default:
                  summary: Example success response
                  value:
                    success: true
                    data:
                    message: Operation completed successfully
        400:
          description: Bad request
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  error:
                    type: string
                  message:
                    type: string
                  details:
                    type: array
                    items:
                      type: string
              examples:
                default:
                  summary: Example error response
                  value:
                    success: false
                    error: Bad Request
                    message: Invalid request parameters
                    details:
                      - Field validation failed
        500:
          description: Internal server error
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  error:
                    type: string
                  message:
                    type: string
              examples:
                default:
                  summary: Example server error response
                  value:
                    success: false
                    error: Internal Server Error
                    message: An unexpected error occurred
      parameters:
        -
          name: limit
          in: query
          description: Maximum number of stocks
          required: false
          schema:
            type: string
  /api/leaderboards:
    get:
      summary: Get all market leaderboards (gainers, losers, most active)
      description: "Implements Requirement 18.5: Display all leaderboards Query Parameters: - limit: Maximum number of stocks per leaderboard (optional, default: 10, max: 50) (实现需求: 18.5)"
      tags:
        - Market
      responses:
        200:
          description: Successful response
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
                  message:
                    type: string
              examples:
                default:
                  summary: Example success response
                  value:
                    success: true
                    data:
                    message: Operation completed successfully
        400:
          description: Bad request
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  error:
                    type: string
                  message:
                    type: string
                  details:
                    type: array
                    items:
                      type: string
              examples:
                default:
                  summary: Example error response
                  value:
                    success: false
                    error: Bad Request
                    message: Invalid request parameters
                    details:
                      - Field validation failed
        500:
          description: Internal server error
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  error:
                    type: string
                  message:
                    type: string
              examples:
                default:
                  summary: Example server error response
                  value:
                    success: false
                    error: Internal Server Error
                    message: An unexpected error occurred
      parameters:
        -
          name: limit
          in: query
          description: Maximum number of stocks per leaderboard
          required: false
          schema:
            type: string
  /api/overview:
    get:
      summary: Get complete market overview (indices, sentiment, leaderboards)
      description: "Combines all market data endpoints Query Parameters: - limit: Maximum number of stocks per leaderboard (optional, default: 10, max: 50)"
      tags:
        - Market
      responses:
        200:
          description: Successful response
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
                  message:
                    type: string
              examples:
                default:
                  summary: Example success response
                  value:
                    success: true
                    data:
                    message: Operation completed successfully
        400:
          description: Bad request
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  error:
                    type: string
                  message:
                    type: string
                  details:
                    type: array
                    items:
                      type: string
              examples:
                default:
                  summary: Example error response
                  value:
                    success: false
                    error: Bad Request
                    message: Invalid request parameters
                    details:
                      - Field validation failed
        500:
          description: Internal server error
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  error:
                    type: string
                  message:
                    type: string
              examples:
                default:
                  summary: Example server error response
                  value:
                    success: false
                    error: Internal Server Error
                    message: An unexpected error occurred
      parameters:
        -
          name: limit
          in: query
          description: Maximum number of stocks per leaderboard
          required: false
          schema:
            type: string
...
components:
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...